Refactor: Move key event handler to SearchBar
